lsn 7

+++ app/Http/Controllers/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\PostRequest;
 use App\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -9,6 +10,11 @@ use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -48,13 +54,13 @@ class PostController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(Request $request)
+    public function store(PostRequest $request)
     {
         $post = new Post();
         $post->title = $request->title;
         $post->short_title = Str::length($request->title)>30 ? Str::substr($request->title. 0, 30) . '...' : $request->title;
         $post->desc = $request->desc;
-        $post->author_id = rand(1, 4);
+        $post->author_id = \Auth::user()->id;
 
         if ($request->file('img')) {
             $path = Storage::putFile('public', $request->file('img'));
@@ -71,12 +77,16 @@ class PostController extends Controller
      * Display the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Http\Response|\Illuminate\View\View
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function show($id)
     {
         $post = Post::join('users', 'author_id', '=', 'users.id')
             ->find($id);
+        if(!$post) {
+            return redirect()->route('post.index')->withErrors('Тебе здесь делать нечего');
+        }
+
         return view('posts.show', compact('post'));
     }
 
@@ -84,11 +94,20 @@ class PostController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Http\Response|\Illuminate\View\View
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(!$post) {
+            return redirect()->route('post.index')->withErrors('Тебе здесь делать нечего');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост');
+        }
+
         return view('posts.edit', compact('post'));
     }
 
@@ -99,9 +118,17 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(Request $request, $id)
+    public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+
+        if(!$post) {
+            return redirect()->route('post.index')->withErrors('Тебе здесь делать нечего');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост');
+        }
         $post->title = $request->title;
         $post->short_title = Str::length($request->title)>30 ? Str::substr($request->title. 0, 30) . '...' : $request->title;
         $post->desc = $request->desc;
@@ -126,6 +153,14 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(!$post) {
+            return redirect()->route('post.index')->withErrors('Тебе здесь делать нечего');
+        }
+
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете удалить данный пост');
+        }
         $post->delete();
         return redirect()->route('post.index', compact('id'))->with('success', 'Пост успешно удален');
     }
